backend/requirements.txt
pymongo==3.12.0
pydantic==1.8.2
python-dotenv==0.19.1
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
